.build/build.ps1: exit if `cmake` has a non-zero exit code
Until now, if the project failed to build, then `cmake --build` exited
with code 1 but the PowerShell script would continue after that (only
the step `cp Debug\kaitai_struct_cpp_stl_runtime.dll tests\Debug` would
fail due to the absence of `Debug\kaitai_struct_cpp_stl_runtime.dll`).
This is not what we want - there is no point to continue running if any
of the `cmake` invocations fail, i.e. return a non-zero exit code.

We want the same behavior as in the POSIX shell script (`.build/build`),
which would terminate immediately after any of the command exited with a
non-zero code due to `set -e`. The equivalent in PowerShell is
apparently `$PSNativeCommandUseErrorActionPreference = $true`, see
https://learn.microsoft.com/en-us/powershell/scripting/learn/experimental-features?view=powershell-7.4#psnativecommanderroractionpreference.
However, this feature is still considered experimental, and it is only
available since PowerShell 7.3 (see
https://learn.microsoft.com/en-us/powershell/scripting/learn/experimental-features?view=powershell-7.4#available-features).

Since we only have two `cmake` invocations, I decided to explicitly
check the exit code of each instead of relying on this experimental
feature.