bytes_to_str() Win32: explicitly list `UTF-32*` as unsupported
I think it is better to admit that they're not implemented yet than to
pretend they don't exist. Explicitly returning `KAITAI_CP_UNSUPPORTED`
serves like a "TODO" comment, while omitting them pretends everything is
fine. They are also in KSC's
[`EncodingList`](https://github.com/kaitai-io/kaitai_struct_compiler/blob/5832a81a48e10c3c207748486e09bd58b9aa4000/shared/src/main/scala/io/kaitai/struct/EncodingList.scala#L21-L22),
so the fact that the `KS_STR_ENCODING_WIN32API` mode doesn't support
them is a limitation and we should be able to see it directly in the
code.