Make the `KS_ZLIB` flag public
When using the runtime library with zlib support via CMake, the required
`KS_ZLIB` flag was not propagated to the linking target before this
commit because it had the `PRIVATE` scope. As a result, the
`kaitai/kaitaistream.h` header file did not have the `KS_ZLIB` macro
defined when included from an external project, and thus did not declare
the `process_zlib` method.