to_string(): define `long long` overloads even in C++98 mode
Since `uint64_t` is an alias for `unsigned long long` on macOS, I think
this is the only way (except for templates that we used until
https://github.com/kaitai-io/kaitai_struct_cpp_stl_runtime/pull/76) to
ensure that `to_string()` supports an `uint64_t` argument.

`long long` is technically not in C++98, but if we detect that the
`LLONG_MAX` macro is defined, we should be able to use it. And as I
mentioned in `Common.cmake`, I don't think we currently support any
compiler without `long long` support, given that we already require
`int64_t`/`uint64_t`, which are also only part of C++11 and not C++98.